update global/system realms handling

+++ src/main/java/it/smartcommunitylab/aac/core/service/RealmService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import org.springframework.beans.factory.InitializingBean;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.support.PageableExecutionUtils;
@@ -12,23 +13,68 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 import org.springframework.util.StringUtils;
 
+import it.smartcommunitylab.aac.SystemKeys;
 import it.smartcommunitylab.aac.common.AlreadyRegisteredException;
 import it.smartcommunitylab.aac.common.NoSuchRealmException;
+import it.smartcommunitylab.aac.common.RegistrationException;
+import it.smartcommunitylab.aac.common.SystemException;
 import it.smartcommunitylab.aac.core.persistence.RealmEntity;
 import it.smartcommunitylab.aac.core.persistence.RealmEntityRepository;
 import it.smartcommunitylab.aac.model.Realm;
 
 @Service
-public class RealmService {
+public class RealmService implements InitializingBean {
+    
+    public static final Set<String> RESERVED_SLUG;
+
+    static {
+        Set<String> s = new HashSet<>();
+        s.add(SystemKeys.REALM_SYSTEM);
+        s.add(SystemKeys.REALM_COMMON);
+        s.add(SystemKeys.REALM_INTERNAL);
+        s.add("aac");
+        RESERVED_SLUG = s;
+    }
 
     private final RealmEntityRepository realmRepository;
+
+    // static immutable systemRealm
+    private final Realm systemRealm;
+    
 
     public RealmService(RealmEntityRepository realmRepository) {
         Assert.notNull(realmRepository, "realm repository is mandatory");
         this.realmRepository = realmRepository;
+        
+        //build system realm
+        systemRealm = new Realm(SystemKeys.REALM_SYSTEM, SystemKeys.REALM_SYSTEM);
+        systemRealm.setEditable(false);
+        systemRealm.setPublic(false);
+        
+    }
+
+    @Override
+    public void afterPropertiesSet() throws Exception {
+        Assert.notNull(systemRealm, "system realm can not be null");
     }
 
     public Realm addRealm(String slug, String name) throws AlreadyRegisteredException {
+        if(!StringUtils.hasText(slug)) {
+            throw new RegistrationException("a valid slug is required");
+        }
+        
+        if (SystemKeys.REALM_GLOBAL.equals(slug) || SystemKeys.REALM_SYSTEM.equals(slug)) {
+            throw new IllegalArgumentException("system realms are immutable");
+        }
+        
+        if(RESERVED_SLUG.contains(slug)) {
+            throw new RegistrationException("slug is reserved");
+        }
+            
+        if(!StringUtils.hasText(name)) {
+            name = slug;
+        }
+        
         RealmEntity r = realmRepository.findBySlug(slug);
         if (r != null) {
             throw new AlreadyRegisteredException("slug already exists");
@@ -45,6 +91,14 @@ public class RealmService {
     }
 
     public Realm findRealm(String slug) {
+        if(!StringUtils.hasText(slug)) {
+            return null;
+        }
+        
+        if (SystemKeys.REALM_SYSTEM.equals(slug)) {
+           return systemRealm;
+        }
+        
         RealmEntity r = realmRepository.findBySlug(slug);
         if (r == null) {
             return null;
@@ -54,21 +108,27 @@ public class RealmService {
     }
 
     public Realm getRealm(String slug) throws NoSuchRealmException {
-        RealmEntity r = realmRepository.findBySlug(slug);
-        if (r == null) {
+       Realm realm = findRealm(slug);
+        if (realm == null) {
             throw new NoSuchRealmException();
         }
 
-        return toRealm(r);
+        return realm;
     }
 
-    public Realm updateRealm(String slug, String name) throws NoSuchRealmException {
+    public Realm updateRealm(String slug, String name, boolean isEditable, boolean isPublic) throws NoSuchRealmException {
+        if (SystemKeys.REALM_GLOBAL.equals(slug) || SystemKeys.REALM_SYSTEM.equals(slug)) {
+            throw new IllegalArgumentException("system realms are immutable");
+        }
+        
         RealmEntity r = realmRepository.findBySlug(slug);
         if (r == null) {
             throw new NoSuchRealmException();
         }
 
         r.setName(name);
+        r.setEditable(isEditable);
+        r.setPublic(isPublic);
 
         r = realmRepository.save(r);
 
@@ -77,12 +137,17 @@ public class RealmService {
     }
 
     public void deleteRealm(String slug) {
+        if (SystemKeys.REALM_GLOBAL.equals(slug) || SystemKeys.REALM_SYSTEM.equals(slug)) {
+            throw new IllegalArgumentException("system realms are immutable");
+        }
+        
         RealmEntity r = realmRepository.findBySlug(slug);
         if (r != null) {
             realmRepository.delete(r);
         }
     }
 
+    
     public List<Realm> listRealms() {
         List<RealmEntity> realms = realmRepository.findAll();
         return realms.stream().map(r -> toRealm(r)).collect(Collectors.toList());
@@ -95,12 +160,14 @@ public class RealmService {
 
         return realms.stream().map(r -> toRealm(r)).collect(Collectors.toList());
     }
+
     public Page<Realm> searchRealms(String keywords, Pageable pageRequest) {
-    	Page<RealmEntity> page = StringUtils.hasText(keywords) ? realmRepository.findByKeywords(keywords, pageRequest) : realmRepository.findAll(pageRequest);
-    	return PageableExecutionUtils.getPage(
-    			page.getContent().stream().map(r -> toRealm(r)).collect(Collectors.toList()),
-    			pageRequest,
-    			() -> page.getTotalElements());
+        Page<RealmEntity> page = StringUtils.hasText(keywords) ? realmRepository.findByKeywords(keywords, pageRequest)
+                : realmRepository.findAll(pageRequest);
+        return PageableExecutionUtils.getPage(
+                page.getContent().stream().map(r -> toRealm(r)).collect(Collectors.toList()),
+                pageRequest,
+                () -> page.getTotalElements());
     }
 
     /*
@@ -110,4 +177,4 @@ public class RealmService {
         Realm r = new Realm(re.getSlug(), re.getName());
         return r;
     }
-}
+}}
